style: fix the eslint warnings
